Time-of-Use/Predictions: improvements (#1028)
- make `TimeOfUsePrices` and `Prediction` more transparent/obvious to use, by using a map of ZonedDateTime to value/price
  - this also allows to handle more than 24h of data if available, e.g. after 2pm when there is data available till next day evening.
   - This is useful, e.g. for ESS-TimeOfUse-Controller, to find better Schedules.
- use Double instead of Float for prices, because it is better usable in Java.
- set Thread priority for Optimizer to low (`MIN_PRIORITY`), to not disturb real-time control
- drop unused JSONRPC-Request
- Persistence-Model-Predictor predicts 48 hours (= always more then prices available)
- Handle non-available Production prediction (for systems without PV)

Reviewed-on: https://git.intranet.fenecon.de/FENECON/fems/pulls/1028

+++ io.openems.edge.controller.ess.timeofusetariff/src/io/openems/edge/controller/ess/timeofusetariff/optimizer/Period.java
@@ -11,7 +11,7 @@ import java.util.regex.Pattern;
 import io.openems.edge.controller.ess.timeofusetariff.StateMachine;
 
 public record Period(ZonedDateTime time, int production, int consumption, int essInitial, int essMaxCharge,
-		int essMaxDischarge, StateMachine state, int essChargeDischarge, int grid, float price, double cost) {
+		int essMaxDischarge, StateMachine state, int essChargeDischarge, int grid, double price, double cost) {
 
 	private static final DateTimeFormatter FORMATTER = DateTimeFormatter.ofPattern("HH:mm");
 
@@ -72,7 +72,7 @@ public record Period(ZonedDateTime time, int production, int consumption, int es
 				StateMachine.valueOf(matcher.group("state")), //
 				Integer.parseInt(matcher.group("essChargeDischarge")), //
 				Integer.parseInt(matcher.group("grid")), //
-				Float.parseFloat(matcher.group("price")), //
+				Double.parseDouble(matcher.group("price")), //
 				Double.parseDouble(matcher.group("cost")));
 	}
 }
